fix: update invitation api
