+++ src/resources/auth/index.ts
@@ -5,6 +5,5 @@ export * from './space-roles';
 export * from './space-role-users';
 export * from './credentials';
 export * from './users';
-export * from './invitation';
 export * from './join-space';
 export * from './presigned-url';
